requirements.txt
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-core>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0

pydantic>=2.7,<3

requests>=2.31.0
tenacity>=8.2.0
numpy>=1.24.0

python-dotenv>=1.0.0

jupyter>=1.0.0
ipykernel>=6.0.0
